pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scramble"
version = "0.1.0"
authors = [{ name = "lux" }]
description = "Semantic Compression for AI Dialogue"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "rich",
    "nltk",
    "anthropic",
    "click",
    "sentence-transformers>=2.5.0",
    "numpy<2",
    "dateparser",
    "textual",
    "pyte",
    "sqlite-vec>=0.1.6",    # SQLite vector search extension
    "llama-index-core>=0.12.0",  
    "llama-index-llms-ollama", #local model for llamaindex
    "duckdb",
    "httpx>=0.26.0",
    "fastapi>=0.109.0",
    "ollama>=0.1.6",
    "pydantic>=2.5.3",
    "torch",
    "pytest",
    "aiofiles",
]

[project.scripts]
ramble = "ramble.app:cli"
maxx = "ramblemaxx.app:RambleMaxx.run" 

[tool.hatch.build.targets.wheel]
packages = ["scramble", "ramble", "ramblemaxx"]

[tool.hatch.metadata]
allow-direct-references = true